chore: bump chart-testing timeout to 20m in CI
The SSH Portal API attempts to connect to Keycloak and exits with an
error if it fails. Kubernetes Crashloop backoff timeout is capped at 5m.

Keycloak can take 15m to start which means that the total time for all
pods to become ready can be higher than 15m.

Therefore bump this timeout in the default chart-testing configuration.